Replace old notefield with new one
