+++ OpenChart/src/UI/NoteField/DrawingContext.cs
@@ -1,6 +1,6 @@
 using OpenChart.Charting.Properties;
 
-namespace OpenChart.UI.NoteField.OpenGL
+namespace OpenChart.UI.NoteField
 {
     public class DrawingContext
     {
